pyproject.toml
[project]
name = "document-search"
version = "0.1.0"
description = "Semantic document search system with ChromaDB and MCP server for Claude Code"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.2",
    "mcp>=0.9.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
